Add dev docs for metrics
As a developer, it's hard to visualize how the statsd metrics are
related to the sequence of actions taken during pipeline processing.

This adds a developer documentation page with a table that shows the
pipeline processing steps, then groups them together and shows which
metrics represent those steps.  It also indicates where timestamps
are stored on objects to form the basis of elapsed times in the
metrics.

Change-Id: I7180bdfb73e14cc67718ecdc6446edb5a24d00ed