zuul_console: namespace console log file
When running Zuul on a long-lived host, nothing clears out the
/tmp/console-log-* files.  We can use a cron job to do this, but
"console*" is a bit of a worryingly wide glob to match.

Namespacing this with a "zuul-" prefix increases confidence you won't
accidentally delete something else important, but generally is just a
bit more informative about where the file comes from.

Change-Id: I199313e1fff567b176020ced1d8e92c5059012cd