zuul-stream: start new console for stream tests
The zuul-stream-functional tests currently run Ansible against two
hosts (node1 and node2) started by the "infrastructure" Zuul.  However
the tests are working with the zuul_console running on the default
port 19885 -- which is the zuul_console started by the setup jobs at
[1].

The result of this is that this test is talking to a zuul_console
instance started from the current executor's code, not the Zuul
checkout uder test -- i.e. changes to zuul_console.py aren't tested by
this change, which I don't think is the intent.

This modifies the job to start its own zuul_console service on the two
hosts running at another port (19887) and updates the environment
variable to talk to this.  This way we are testing against the
zuul_console code from the Zuul checkout.

[1] https://opendev.org/zuul/zuul-jobs/src/branch/master/roles/prepare-workspace/tasks/main.yaml#L2

Change-Id: Ib11b77cfdc6c59d12807c6d9684c3e653ccad863