zuul_console: cleanup console files
A refresher on how this works, to the best of my knowledge

 1 Firstly, Zuul's Ansible has a library task "zuul_console:" which is
   run against the remote node; this forks a console daemon, listening
   on a default port.

 2 We have a action plugin that runs for each task, and if that task
   is a command/shell task, assigns it a unique id

 3 We then override with library/command.py (which backs command/shell
   tasks) with a version that forks and runs the process on the target
   node as usual, but also saves the stdout/stderr output to a
   temporary file named per the unique uuid from the step above.

 4 At the same time we have the callback plugin zuul_stream.py, which
   Ansible is calling as it moves through starting, running and
   finishing the tasks.  This looks at the task, and if it has a UUID
   [2], sends a request to the zuul_console [1], which opens the
   temporary file [3] and starts streaming it back.

 5 We loop reading this until the connecion is closed by [1],
   eventaully outputting each line.

In this way, the console log is effectively streamed and saved into
our job output.

One issue with this is that the temporary file from [3] is not cleared
up for a long-running zuul_console.

This proposes a small protocol between the callback and console, where
the streamer requests the file to be streamed with "s:<uuid>" but then
also indicates when it is finished with a "f:<uuid>".  At this point,
the zuul_console dameon can remove the temporary file.

If there is no prefix, it just assumes it is the raw UUID, thus
maintaining backwards compatibility.

To test this we look at the console output of Ansible in the stream
test, which has a log giving the UUID of task runs.  We then list out
the temporary files on the test nodes, and check that this UUID
doesn't exist; meaning the file has been cleaned up.

Change-Id: I823156dc2bcae91bd6d9770bd1520aa55ad875b4