Update patch set 10
Patch Set 10:

For what it's worth, this was an intentional design decision, not an accident (doesn't mean we can't revisit it, but just providing context).  Partly for simplicity of the remote code, and partly for the following.
 
This will always be subject to a DoS since there's basically no access control on zuul_console.  An attacker would be able to remove the streaming logs of an attack in progress if they can access the streaming port.

Patch-set: 10
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>