Do not deduplicate the noop job
There is no good reason to do so (there are no resources consumed
by the job), and it's difficult to disable a behavior for the
noop job globally since it has no definition.  Let's never have it
deduplicate so that we keep things simple for folks who want to
avoid deduplication.

Change-Id: Ib3841ce5ef020540edef1cfa479d90c65be97112