Update patch set 2
Patch Set 2:

(2 comments)

I think I understand why you want this (the commit message explains that), but I'm not sure we should expose this in builds; I think it's potentially misleading.  I'd rather find another way to accomplish what you want, either by enhancements to Nodepool so that it can satisfy requests for resources (that's the part of Zuul that is responsible for that).  Or maybe do something with hitting the Zuul REST API so you can calculate your own priority value, but it reflects the current state of the system, and not frozen at some arbitrary point in time.

Patch-set: 2
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>