WIP: Fix deduplicated build references
When we deduplicate a build, we use the same Build object in
multiple BuildSets.  But when a second scheduler loads the pipeline
state, even if the build path points to a build outside of a given
BuildSet it will create new Build objects and set their local object
references for the buildset and job to the "current" buildset rather
than the one that is being referenced.  Instead we need to find the
Build object for the other buildset, but at this point in the process,
the queues haven't been completely loaded yet and so we can't go
searching for it.

To correct this, when we load a Build from ZooKeeper that is outside
of our buildset, we insert a BuildReference object which is a
placeholder until we finish loading the pipeline state.  Then at the
end of that process, we go through and replace the BuildReference
objects with the actual builds.

Change-Id: I0f0e5e4b8f8c9d34b86ccd83ad0b0578ce4a780c