Fix duplicate setResult calls in deduplicated builds
We call item.setResult after a build is complete so that the queue
item can do any internal processing necessary (for example, prepare
data structures for child jobs, or move the build to the retry_builds
list).

In the case of deduplicated builds, we should do that for every queue
item the build participates in since each item may have a different
job graph.

We were not correctly identifying other builds of deduplicated jobs
and so in the case of a dependency cycle we would call setResult on
jobs of the same name in that cycle regardless of whether they were
deduplicated.

This corrects the issue and adds a test to detect that case.

Change-Id: I4c47beb2709a77c21c11c97f1d1a8f743d4bf5eb