WIP: Alternative deduplication implementation
This is an exploration of an alternative implementation for job
deduplication.  Rather than using the same Build objects in multiple
BuildSets, this copies the Build object in the other buildsets.

This mean that each Build object can be read and written from ZK
directly.  But it also means there are multiple Build objects with
the same UUID.  It also stores multiple copies of the result data
in ZK and when we receive a result event, we call the result update
methods on every copy of the build.  Since the SQL database schema
can't handle more than one build with the same UUID, we store a flag
indicating whether a given Build is the original or a copy, and we
only update the build table in the DB for the original.

If we decide we like this better than the current approach of direct
object reuse, then we will need to come up with a transition plan
(we will probably need to update the code to support both approaches
depending on what is currently written in ZK and switch to writing
this approach on a schema version bump).

This also does not pass tests yet, and may need significant additional
work.

Change-Id: I7625b4d824dce99b17a1dc7a1ef5c7947c455a64