zuul-stream : don't write out logfile for tasks in loops
In zuul_stream.py:v2_playbook_on_task_start() it checks for
"task.loop" and exits if the task is part of a loop.

However the library/command.py override still writes out the console
log despite it never being read.  To avoid leaving this file around,
mark a sentinel uuid in the action plugin if the command is part of a
loop.  In that case, for simplicity we just write to /dev/null -- that
way no other assumptions in the library/command.py have to change; it
just doesn't leave a file on disk.

Change-Id: I4f4660c3c0b0f170561c14940cc159dc43eadc79