zuul-stream : don't write out logfile for tasks in loops
In zuul_stream.py:v2_playbook_on_task_start() it checks for
"task.loop" and exits if the task is part of a loop.

However the library/command.py override still writes out the console
log despite it never being read.  To avoid leaving this file around,
mark a sentinel uuid in the action plugin if the command is part of a
loop.  In that case, for simplicity we just write to /dev/null -- that
way no other assumptions in the library/command.py have to change; it
just doesn't leave a file on disk.

This is currently difficult to test as the infrastructure zuul_console
leaves /tmp/console-* files and we do not know what comes from that,
or testing.  After this and the related change
I823156dc2bcae91bd6d9770bd1520aa55ad875b4 are deployed to the
infrastructure executors, we can make a simple and complete test for
the future by just ensuring no /tmp/console-* files are left behind
afer testing.  I have tested this locally and do not see files from
loops, which I was before this change.

Change-Id: I4f4660c3c0b0f170561c14940cc159dc43eadc79