Reconcile pipeline change cache at start of refresh
The pipeline change cache is used to avoid repeated cache lookups
for change dependencies, but is not as robust as the "real" sounce
change cache at managing the lifetime of change objects.  If it is
used to store some change objects in one scheduler which are later
dequeued by a second scheduler, the next time those changes show
up in that pipeline on the first scheduler it may use old change
objects instead of new ones from the ZooKeeper cache.

To correct this, validate all the pipeline cache entries against the
source cache entries to make sure they are the same object.  We
could just clear the pipeline cache, but we would need to hit ZK
a lot to refresh it (since a ChangeCache.get() involves at least
a quick ZK check to make sure we have the latest data).  This keeps
the pipeline cache pointed to the right objects and as populated
as possible.

Change-Id: Ief968d3358e1c7c4c0723edba3e53c2c2e663132