Use internal links in job graph display
Since we're using the graphviz library to render an svg, we can't
just pass in a react router Link object for the node links we
create for jobs.  But we can manipulate the DOM after the SVG is
produced and add onClick handlers that do the same thing.

This avoids refreshing the page/app when clicking on links in the
graph.

Change-Id: Ie173e3d2e31efd47dbe4a44fda8b743250729339