Web: disable scroll wheel zoom in job graph
When scrolling the project page with a job graph rendered, users
using scroll wheels (or similar input devices) may be surprised
that once their pointer moves over the job graph, it begin zooming
instead of the page scrolling.  Disable zooming the job graph
with the mouse wheel so that the behavior remains consistent.

Zooming is still possible with click/shift-click.

Change-Id: Ic3884f1317cb7799e4e387aa51c81c080e028d10