Use git cli for setting repo remote URL
We see cases where the remote URL is updated on the repo object, but the
git config still contains an old URL with an outdated Github access
token.

It seems that GitPython might swallow some exceptions [0] when writing
the config. To make this more stable we'll use the Git CLI to set the
remote URL. The `remote_url` on the repo object is only updated when the
command was successful.

[0]: https://github.com/gitpython-developers/GitPython/blob/f0c6e1164f390081a27de952552aa83d34035f2a/git/config.py#L414

Change-Id: I2e6535213286be2e4ba3963ecd13e7c47a682220