Set remote URL when different in config
We see cases where the remote URL is updated on the repo object, but the
git config still contains an old URL with an outdated Github access
token.

It seems that GitPython might swallow some exceptions [0] when writing
the config. To avoid issues with outdated Github access tokens in the
URL we check the new URL directly against the Git config.

This also adds a missing repo lock when building the repo state.

[0]: https://github.com/gitpython-developers/GitPython/blob/f0c6e1164f390081a27de952552aa83d34035f2a/git/config.py#L414

Change-Id: I8e1b5b26f03cb75727d2b2e3c9310214a3eac447