Only update repo remote URL after writing config
We see cases where the remote URL is updated on the repo object, but the
git config still contains an old URL with an outdated access token.

This is a first stab at fixing the problem, but it might not solve the
issue as GitPython is swallowing some exceptions [0] when writing the
config.

[0]: https://github.com/gitpython-developers/GitPython/blob/f0c6e1164f390081a27de952552aa83d34035f2a/git/config.py#L414

Change-Id: I2e6535213286be2e4ba3963ecd13e7c47a682220