Force set repo remote URL
We see cases where the remote URL is updated on the repo object, but the
git config still contains an old URL with an outdated Github access
token.

It seems that GitPython might swallow some exceptions [0] when writing
the config. To avoid issues with outdated Github access tokens in the
URL we force-set the remote URL each time.

[0]: https://github.com/gitpython-developers/GitPython/blob/f0c6e1164f390081a27de952552aa83d34035f2a/git/config.py#L414

Change-Id: I8e1b5b26f03cb75727d2b2e3c9310214a3eac447