zuul_console: do not use f-strings
In Ief366c092e05fb88351782f6d9cd280bfae96237 I missed that this runs
in the context of the remote node; meaning that it must support all
the Python versions that might run there.  f-strings are not 3.5
compatible.

I'm thinking about how to lint this better (a syntax check run?)

Change-Id: Ia4133b061800791196cd631f2e6836cb77347664