test-setup-docker: wrap call in env
In weird things about bash today:

 $ cat test.sh
 SUDO=''
 FOO=1 echo "hello"
 ${SUDO} FOO=1 echo "hello"

 $ bash ./test.sh
 hello
 ./test.sh: line 6: FOO=1: command not found

I guess the "${SUDO}" gets tokenized first, doesn't have an "=" in it
so stops processing of that line, and then when bash comes back to it,
it thinks FOO=1 is a command to execute ([1] if you're feeling
masochistic).

Iac32799425a5bd4b1bdbf56f34a2310241ac4499 introduced a sudo wrapper
here that has made this fail with

 + USER_ID=1000 docker-compose up -d
 ./tools/test-setup-docker.sh: line 50: USER_ID=1000: command not found

for me.  Wrap this in "env" to get the desired behaviour.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_10_02

Change-Id: I1126c9cd2e79313f167d3b894ba29cbabde5e24c