zuul_stream: handle non-string msg value
I noticed in some of our testing a construct like

 debug:
   msg: '{{ ansible_version }}'

was actually erroring out; you'll see in the console output if you're
looking

 Ansible output: b'TASK [Print ansible version msg={{ ansible_version }}] *************************'
 Ansible output: b'[WARNING]: Failure using method (v2_runner_on_ok) in callback plugin'
 Ansible output: b'(<ansible.plugins.callback.zuul_stream.CallbackModule object at'
 Ansible output: b"0x7f502760b490>): 'dict' object has no attribute 'startswith'"

and the job-output.txt will be empty.

This is because the msg value here comes in as a dict, and in several
places we assume it is a string.  This changes places we inspect the
msg variable to use the standard Ansible way to make a text string
(to_text function) and ensures in the logging function it converts the
input to a string.

The testing added maybe doesn't feel like the optimal place for this,
but it should stop regressions in printing this directly and in a loop
(which take different paths in the plugin).

Change-Id: I6e6ed8dba2ba1fc74e7fc8361e8439ea6139279e