zuul-tox-remote: capture callback errors in remote zuul jobs
We have a couple of places that can trigger errors in the zuul_stream
callback that this testing currently misses.

To try and catch this case better we grab the Ansible console output
(where the failure of the callback plugin is noted) and check it for
Ansible failure strings.

Just for review purposes, follow-on changes will correct current
errors so this test can be inverted.

Change-Id: I9f569a411729f8a067de17d99ef6b9d74fc21543