zuul_stream: handle failed_when tasks
Currently the task in the test playbook

  - hosts: compute1
    tasks:
    - name: Command Not Found
      command: command-not-found
      failed_when: false

is failing in the zuul_stream callback with an exception trying to
fill out the "delta" value in the message here.  The result dict
(taken from the new output) shows us why:

 2022-08-24 07:19:27.079961 | TASK [Command Not Found]
 2022-08-24 07:19:28.578380 | compute1 | ok: ERROR (ignored)
 2022-08-24 07:19:28.578622 | compute1 | {
 2022-08-24 07:19:28.578672 | compute1 |   "failed_when_result": false,
 2022-08-24 07:19:28.578700 | compute1 |   "msg": "[Errno 2] No such file or directory: b'command-not-found'",
 2022-08-24 07:19:28.578726 | compute1 |   "rc": 2
 2022-08-24 07:19:28.578750 | compute1 | }

i.e. it has no start/stop/delta in the result (it did run and fail, so
you'd think it might ... but this is what Ansible gives us).

This checks for this path; as mentioned the output will now look like
above in this case.

This was found by the prior change I9f569a411729f8a067de17d99ef6b9d74fc21543

Change-Id: I106b2bbe626ed5af8ca739d354ba41eca2f08f77